+++ src/index/rtx.rs
@@ -9,8 +9,8 @@ impl Rtx<'_> {
         .0
         .open_table(HEIGHT_TO_BLOCK_HASH)?
         .range(0..)?
-        .rev()
-        .next()
+        .next_back()
+        .and_then(|result| result.ok())
         .map(|(height, _hash)| Height(height.value())),
     )
   }
@@ -21,8 +21,8 @@ impl Rtx<'_> {
         .0
         .open_table(HEIGHT_TO_BLOCK_HASH)?
         .range(0..)?
-        .rev()
-        .next()
+        .next_back()
+        .and_then(|result| result.ok())
         .map(|(height, _hash)| height.value() + 1)
         .unwrap_or(0),
     )
@@ -42,8 +42,8 @@ impl Rtx<'_> {
           .0
           .open_table(HEIGHT_TO_BLOCK_HASH)?
           .range(0..)?
-          .rev()
-          .next()
+          .next_back()
+          .and_then(|result| result.ok())
           .map(|(_height, hash)| BlockHash::load(*hash.value())),
       ),
     }